refactor(api): move authentication method repository
from shared to identity access management
