+++ api/src/identity-access-management/domain/usecases/index.js
@@ -12,10 +12,10 @@ import * as userService from '../../../shared/domain/services/user-service.js';
 import * as passwordValidator from '../../../shared/domain/validators/password-validator.js';
 import * as userValidator from '../../../shared/domain/validators/user-validator.js';
 import * as adminMemberRepository from '../../../shared/infrastructure/repositories/admin-member-repository.js';
-import * as authenticationMethodRepository from '../../../shared/infrastructure/repositories/authentication-method-repository.js';
 import * as userLoginRepository from '../../../shared/infrastructure/repositories/user-login-repository.js';
 import { injectDependencies } from '../../../shared/infrastructure/utils/dependency-injection.js';
 import { importNamedExportsFromDirectory } from '../../../shared/infrastructure/utils/import-named-exports-from-directory.js';
+import * as authenticationMethodRepository from '../../infrastructure/repositories/authentication-method.repository.js';
 import { oidcProviderRepository } from '../../infrastructure/repositories/oidc-provider-repository.js';
 import * as resetPasswordDemandRepository from '../../infrastructure/repositories/reset-password-demand.repository.js';
 import * as userRepository from '../../infrastructure/repositories/user.repository.js';
