+++ api/tests/integration/domain/usecases/add-pix-authentication-method-by-email_test.js
@@ -1,7 +1,7 @@
 import { AuthenticationMethodAlreadyExistsError } from '../../../../lib/domain/errors.js';
 import { addPixAuthenticationMethodByEmail } from '../../../../lib/domain/usecases/add-pix-authentication-method-by-email.js';
+import * as authenticationMethodRepository from '../../../../src/identity-access-management/infrastructure/repositories/authentication-method.repository.js';
 import * as userRepository from '../../../../src/identity-access-management/infrastructure/repositories/user.repository.js';
-import * as authenticationMethodRepository from '../../../../src/shared/infrastructure/repositories/authentication-method-repository.js';
 import { catchErr, databaseBuilder, expect } from '../../../test-helper.js';
 
 describe('Integration | UseCase | add-pix-authentication-method-by-email', function () {
